created components for form on each register, login and forgot-password screen and added lottie animation
