+++ src/features/account/screens/forgot-password.screen.js
@@ -1,16 +1,6 @@
-import React, { useContext } from "react";
-import { ScrollView } from "react-native";
-import {
-  TextInput,
-  HelperText,
-  ActivityIndicator,
-  Colors,
-  Button,
-} from "react-native-paper";
-import GradientButton from "react-native-gradient-buttons";
-import { Formik } from "formik";
+import React from "react";
+import { Button } from "react-native-paper";
 
-import * as yup from "yup";
 import styled from "styled-components/native";
 
 import { colors } from "../../../infrastructure/theme/colors";
@@ -18,29 +8,11 @@ import { SafeArea } from "../../../components/utilities/safe-area.components";
 import { Spacer } from "../../../components/spacer/spacer.components";
 import { Text } from "../../../components/typography/text.components";
 import { FacebookAndGoogleButtons } from "../components/fb-and-google-buttons.components";
-
-import { AuthenticationContext } from "../../../services/authentication/authentication.context";
+import { FormForgotPassword } from "../components/form-forgot-password.components";
 
 const TopTextSection = styled.View`
   padding: ${(props) => props.theme.space[4]};
   align-items: center;
-`;
-
-const TextInputsSection = styled.View`
-  padding-top: ${(props) => props.theme.space[3]};
-  padding-left: ${(props) => props.theme.space[3]};
-  padding-right: ${(props) => props.theme.space[3]};
-`;
-
-const SendButtonSection = styled.View`
-  padding-left: ${(props) => props.theme.space[2]};
-  padding-right: ${(props) => props.theme.space[2]};
-`;
-
-const SendText = styled(Text)`
-  color: white;
-  font-family: ${(props) => props.theme.fonts.body_medium};
-  font-size: ${(props) => props.theme.fontSizes.body};
 `;
 
 const ForgotPasswordText = styled(Text)`
@@ -75,30 +47,11 @@ const OrText = styled(Text)`
   width: 50px;
 `;
 
-const validationSchema = yup.object().shape({
-  email: yup
-    .string()
-    .label("Email")
-    .email("Enter a valid email")
-    .required("Please enter a registered email"),
-});
+const WhiteSpaceSection = styled.View`
+  flex-grow: 1;
+`;
 
 export const ForgotPasswordScreen = ({ navigation }) => {
-  const { onPasswordReset, error, isLoading } = useContext(
-    AuthenticationContext
-  );
-
-  const handlePasswordReset = async (values, actions) => {
-    const { email } = values;
-
-    try {
-      onPasswordReset(email);
-      navigation.navigate("Login");
-    } catch (error) {
-      actions.setFieldError("general", error.message);
-    }
-  };
-
   return (
     <SafeArea>
       <TopTextSection>
@@ -109,77 +62,15 @@ export const ForgotPasswordScreen = ({ navigation }) => {
           account.
         </MessageText>
       </TopTextSection>
-      <Formik
-        initialValues={{ email: "" }}
-        onSubmit={(values, actions) => {
-          handlePasswordReset(values, actions);
-        }}
-        validationSchema={validationSchema}
-      >
-        {({
-          handleChange,
-          values,
-          handleSubmit,
-          errors,
-          touched,
-          handleBlur,
-        }) => (
-          <ScrollView scrollEnabled={false}>
-            <TextInputsSection>
-              <TextInput
-                mode="outlined"
-                value={values.email}
-                onChangeText={handleChange("email")}
-                onBlur={handleBlur("email")}
-                label="Enter email"
-                style={{ height: 50 }}
-                theme={{
-                  roundness: 15,
-                  colors: {
-                    primary: colors.brand.primary,
-                    nderlineColor: "blue",
-                    placeholder: "#cecbce",
-                    background: "white",
-                  },
-                }}
-              />
-              <HelperText type="error" visible={errors}>
-                {touched.email && errors.email}
-              </HelperText>
-            </TextInputsSection>
-
-            <SendButtonSection>
-              {!isLoading ? (
-                <GradientButton
-                  text={<SendText>Send</SendText>}
-                  gradientBegin={colors.brand.primary}
-                  gradientEnd="#6dd5ed"
-                  gradientDirection="vertical"
-                  height={50}
-                  radius={15}
-                  onPressAction={handleSubmit}
-                />
-              ) : (
-                <ActivityIndicator animating={true} color={Colors.blue300} />
-              )}
-            </SendButtonSection>
-
-            <Text
-              style={{ color: "red", paddingHorizontal: 40, marginTop: 10 }}
-            >
-              {errors.general}
-            </Text>
-
-            <OrSection>
-              <HorizontalLine />
-              <OrText>OR</OrText>
-              <HorizontalLine />
-            </OrSection>
-            <Spacer size="medium" />
-            <FacebookAndGoogleButtons />
-          </ScrollView>
-        )}
-      </Formik>
+      <FormForgotPassword onNavigate={navigation.navigate} />
+      <OrSection>
+        <HorizontalLine />
+        <OrText>OR</OrText>
+        <HorizontalLine />
+      </OrSection>
+      <Spacer size="medium" />
+      <FacebookAndGoogleButtons />
+      <WhiteSpaceSection />
       <Button
         color="white"
         uppercase={false}
