+++ src/features/account/screens/login.screen.js
@@ -1,16 +1,6 @@
-import React, { useContext } from "react";
-import { ScrollView } from "react-native";
-import {
-  TextInput,
-  HelperText,
-  Button,
-  ActivityIndicator,
-  Colors,
-} from "react-native-paper";
-import GradientButton from "react-native-gradient-buttons";
-import { Formik } from "formik";
+import React from "react";
+import { Button } from "react-native-paper";
 
-import * as yup from "yup";
 import NetInfo from "@react-native-community/netinfo";
 import styled from "styled-components/native";
 
@@ -19,40 +9,16 @@ import { SafeArea } from "../../../components/utilities/safe-area.components";
 import { Spacer } from "../../../components/spacer/spacer.components";
 import { Text } from "../../../components/typography/text.components";
 import { FacebookAndGoogleButtons } from "../components/fb-and-google-buttons.components";
-
-import { AuthenticationContext } from "../../../services/authentication/authentication.context";
+import { FormLogin } from "../components/form-login.components";
 
 const TopTextSection = styled.View`
   padding: ${(props) => props.theme.space[3]};
-`;
-
-const TextInputsSection = styled.View`
-  padding-top: ${(props) => props.theme.space[5]};
-  padding-left: ${(props) => props.theme.space[3]};
-  padding-right: ${(props) => props.theme.space[3]};
-`;
-
-const ForgotPasswordSection = styled.View`
-  align-items: flex-end;
-`;
-
-const LoginButtonSection = styled.View`
-  padding-top: ${(props) => props.theme.space[3]};
-  padding-left: ${(props) => props.theme.space[2]};
-  padding-right: ${(props) => props.theme.space[2]};
-  padding-bottom: ${(props) => props.theme.space[4]};
 `;
 
 const SignUpSection = styled.View`
   align-items: center;
   flex-direction: row;
   justify-content: center;
-`;
-
-const LoginText = styled(Text)`
-  color: white;
-  font-family: ${(props) => props.theme.fonts.body_medium};
-  font-size: ${(props) => props.theme.fontSizes.button};
 `;
 
 const WelcomeText = styled(Text)`
@@ -90,18 +56,9 @@ const OrText = styled(Text)`
   width: 50px;
 `;
 
-const validationSchema = yup.object().shape({
-  email: yup
-    .string()
-    .label("Email")
-    .email("Enter a valid email")
-    .required("Please enter a registered email"),
-  password: yup
-    .string()
-    .label("Password")
-    .required()
-    .min(5, "Password should be at least 5 characters "),
-});
+const WhiteSpaceSection = styled.View`
+  flex-grow: 1;
+`;
 
 // create the Network Information
 const unsubscribe = NetInfo.addEventListener((state) => {
@@ -113,124 +70,21 @@ const unsubscribe = NetInfo.addEventListener((state) => {
 unsubscribe();
 
 export const LoginScreen = ({ navigation }) => {
-  const { onLogin, error, isLoading } = useContext(AuthenticationContext);
-
-  const onSignIn = async (values, actions) => {
-    const { email, password } = values;
-    try {
-      onLogin(email, password);
-    } catch (error) {
-      // this code will hold the error handler to avoid the error message: "the
-      // password is invalid or the user does have a password"
-      actions.setFieldError("general", error.message);
-    } finally {
-      actions.setSubmitting(false);
-    }
-  };
-
   return (
     <SafeArea>
       <TopTextSection>
         <WelcomeText>Welcome</WelcomeText>
         <ContinueLoginText>Log in to continue!</ContinueLoginText>
       </TopTextSection>
-      <Formik
-        initialValues={{ email: "", password: "" }}
-        onSubmit={(values, actions) => onSignIn(values, actions)}
-        validationSchema={validationSchema}
-      >
-        {({
-          handleChange,
-          values,
-          handleSubmit,
-          errors,
-          touched,
-          handleBlur,
-        }) => (
-          <ScrollView scrollEnabled={false}>
-            <TextInputsSection>
-              <TextInput
-                mode="outlined"
-                label="Email"
-                keyboardType="email-address"
-                value={values.email}
-                onChangeText={handleChange("email")}
-                onBlur={handleBlur("email")}
-                style={{ height: 50 }}
-                theme={{
-                  roundness: 15,
-                  colors: {
-                    primary: colors.brand.primary,
-                    nderlineColor: "blue",
-                    placeholder: "#cecbce",
-                    background: "white",
-                  },
-                }}
-              />
-              <HelperText type="error" visible={errors}>
-                {touched.email && errors.email}
-              </HelperText>
-              <TextInput
-                mode="outlined"
-                label="Password"
-                secureTextEntry
-                value={values.password}
-                onChangeText={handleChange("password")}
-                onBlur={handleBlur("password")}
-                style={{ height: 50 }}
-                theme={{
-                  roundness: 15,
-                  colors: {
-                    primary: colors.brand.primary,
-                    nderlineColor: "blue",
-                    placeholder: "#cecbce",
-                    background: "white",
-                  },
-                }}
-              />
-              <HelperText type="error" visible={errors}>
-                {touched.password && errors.password}
-              </HelperText>
-              <ForgotPasswordSection>
-                <Button
-                  color="white"
-                  labelStyle={{
-                    fontSize: 10,
-                    color: colors.brand.primary,
-                  }}
-                  onPress={() => {
-                    navigation.navigate("ForgotPassword");
-                  }}
-                >
-                  Forgot password?
-                </Button>
-              </ForgotPasswordSection>
-            </TextInputsSection>
-            <LoginButtonSection>
-              {!isLoading ? (
-                <GradientButton
-                  text={<LoginText>Login</LoginText>}
-                  gradientBegin={colors.brand.primary}
-                  gradientEnd="#6dd5ed"
-                  gradientDirection="vertical"
-                  height={50}
-                  radius={15}
-                  onPressAction={handleSubmit}
-                />
-              ) : (
-                <ActivityIndicator animating={true} color={Colors.blue300} />
-              )}
-            </LoginButtonSection>
-            <OrSection>
-              <HorizontalLine />
-              <OrText>OR</OrText>
-              <HorizontalLine />
-            </OrSection>
-            <Spacer size="medium" />
-            <FacebookAndGoogleButtons />
-          </ScrollView>
-        )}
-      </Formik>
+      <FormLogin onNavigate={navigation.navigate} />
+      <OrSection>
+        <HorizontalLine />
+        <OrText>OR</OrText>
+        <HorizontalLine />
+      </OrSection>
+      <Spacer size="medium" />
+      <FacebookAndGoogleButtons />
+      <WhiteSpaceSection />
       <SignUpSection>
         <NewUserText>Don't have an account?</NewUserText>
         <Button
